Refactorización del proyecto

+++ src/main/java/com/consorcio/servicios/Repository/ModalityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ModalityRepository extends JpaRepository<Modality, Integer> {
+public interface ModalityRepository extends JpaRepository<Modality, Long> {
 
 }
