+++ src/main/java/com/consorcio/servicios/Repository/ReadingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReadingRepository extends JpaRepository<Reading, Integer> {
+public interface ReadingRepository extends JpaRepository<Reading, Long> {
 
 }
